chat_with_docs/requirements.txt
langchain==0.1.17
langchain-community==0.0.38
langchain-openai==0.1.7
faiss-cpu==1.7.4
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
requests==2.31.0
streamlit==1.29.0
python-dotenv==1.0.0
tiktoken==0.5.2
openai==1.6.1
sentence-transformers==2.2.2
chromadb==0.4.18
